+++ src/FolkerKinzel.Strings/Intls/Utf8Validator.cs
@@ -1,3 +1,4 @@
+using FolkerKinzel.Helpers.Polyfills;
 using FolkerKinzel.Strings.Properties;
 
 namespace FolkerKinzel.Strings.Intls;