+++ src/FolkerKinzel.Strings/StringBuilderExtension_GetPersistentHashCode.cs
@@ -1,4 +1,5 @@
 using System.Security.Cryptography;
+using FolkerKinzel.Helpers.Polyfills;
 using FolkerKinzel.Strings.Intls;
 using FolkerKinzel.Strings.Properties;