+++ src/FolkerKinzel.Strings/StringExtension_ReplaceWhiteSpaceWith.cs
@@ -1,4 +1,5 @@
 using System.Text.RegularExpressions;
+using FolkerKinzel.Helpers.Polyfills;
 using FolkerKinzel.Strings.Intls;
 
 namespace FolkerKinzel.Strings;