+++ src/FolkerKinzel.Strings/StringPolyfillExtension_ReplaceLineEndings.cs
@@ -1,3 +1,4 @@
+using FolkerKinzel.Helpers.Polyfills;
 using FolkerKinzel.Strings.Intls;
 
 namespace FolkerKinzel.Strings;