+++ src/FolkerKinzel.Strings/UrlEncoding.cs
@@ -1,5 +1,6 @@
 using System.Globalization;
 using System.Reflection;
+using FolkerKinzel.Helpers.Polyfills;
 using FolkerKinzel.Strings.Intls;
 
 namespace FolkerKinzel.Strings;