+++ src/FolkerKinzel.Strings/md/ReleaseNotes.md
@@ -1,5 +1,6 @@
-﻿- .NET 9.0 support
-- Solves conflicts that `ReadOnlySpanPolyfillExtension.EndsWith(this ReadOnlySpan<char>, char)` and `ReadOnlySpanPolyfillExtension.StartsWith(this ReadOnlySpan<char>, char)` have with the new extension methods of `System.MemoryExtension` in .NET 9.0.
+﻿- New method overload `int TextEncodingConverter.GetCodePage(string, out int)` that returns the code page of a text file.
+- Dependency update
+- The new dependency `FolkerKinzel.Helpers` helps to avoid code duplication in projects that depend on this library.
 
 &nbsp;
 > **Project reference:** On some systems, the content of the CHM file in the Assets is blocked. Before opening the file right click on the file icon, select Properties, and **check the "Allow" checkbox** - if it is present - in the lower right corner of the General tab in the Properties dialog.